#4: Fix: Unescape Markdown Links (%20 issue)
By default, when pandoc parses a markdown link it also escapes the link as a uri. So [link](foo bar baz) gets parsed as foo%20bar%baz. This is quite good for browser-destined markdown but when we try to resolve it in the filesystem it will be taken literally and fail.