Allow the user to specify which files should be scanned and updated (#20)
This adds --scan-file and --no-scan options to allow the user which files should be scanned and updated as part of whatever operation the user wants to perform. By default md-graph will scan and update links for all files in the library. It only reparses documents that it thinks have changed (based on the file's modification datetime); however in some cases it may be more efficient to update individual documents. For instance you may want to add a vim hook to run md-graph populate -f % and only update the file that has changed.

This enables a future integrate inotify daemon, as well as using md-graph with inotify-wait.